product_management/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
dj-database-url==2.2.0
Django==4.2.16
djangorestframework==3.15.2
dramatiq==1.17.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
